+++ flask_pluginkit/__init__.py
@@ -11,14 +11,16 @@
 
 from .pluginkit import PluginManager
 from .utils import Flask, LocalStorage, RedisStorage
+from ._installer import PluginInstaller
 
 __author__ = "staugur <staugur@saintic.com>"
 
-__version__ = "3.0.1"
+__version__ = "3.1.0"
 
 __all__ = [
     "Flask",
     "PluginManager",
     "LocalStorage",
     "RedisStorage",
+    "PluginInstaller",
 ]